+++ app/src/main/java/com/mateusz113/financemanager/data/repository/dto/PaymentListingDto.kt
@@ -3,14 +3,14 @@ package com.mateusz113.financemanager.data.repository.dto
 class PaymentListingDto() {
     var id: String? = null
     var title: String? = null
-    var amount: Float? = null
+    var amount: Double? = null
     var timestamp: Long? = null
     var category: String? = null
 
     constructor(
         id: String?,
         title: String?,
-        amount: Float?,
+        amount: Double?,
         timestamp: Long?,
         category: String?
     ) : this() {
