+++ app/src/main/java/com/mateusz113/financemanager/domain/model/PaymentDetails.kt
@@ -5,7 +5,7 @@ import java.time.LocalDate
 data class PaymentDetails(
     val title: String,
     val description: String,
-    val amount: Float,
+    val amount: Double,
     val photoUrls: List<String>,
     val date: LocalDate,
     val category: Category
